requirements.txt
discord.py==2.7.1
python-dotenv>=1.0.0,<2.0
requests>=2.31.0,<3.0
